Add hooks to the Go SDK.
Hooks allow a user to tailor behavior of the SDK runtime.
This implementation allows hooks to run at startup, and hook
into the request/response processing of FnAPI messages.

Hooks included in this commit allow for profiling and tracing,
tailoring the behavior of the gRPC transport, and recording
sessions of FnAPI data to be used for playback with the session runner.